Rename instructions.{hpp,cpp} -> instruction_metrics.{hpp,cpp}
This will make space for moving instructions implementations to
instructions.hpp file.
